backend/pyproject.toml
[project]
name = "itonecup-spd"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksey Yakunin",email = "yakunina1111@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "fast-depends (>=2.4.12,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis (>=5.2.1,<6.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "taskiq[reload] (>=0.11.16,<0.12.0)",
    "taskiq-redis (>=1.0.3,<2.0.0)",
    "taskiq-aio-pika (>=0.4.1,<0.5.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "taskiq-pipelines (>=0.1.4,<0.2.0)",
    "openai (>=1.75.0,<2.0.0)",
]

[tool.pyright]
executionEnvironments = [
  { root = "." }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
